feat: create files of type autoencoders
