+++ models/__init__.py
@@ -1,0 +1,4 @@
+from .autoencoder import Autoencoder
+from .convolutional_autoencoder import ConvolutionalAutoencoder
+from .convolutional_vae import ConvolutionalVAE
+from .variational_autoencoder import VariationalAutoencoder
